refactor: switch exercises to MDX
