+++ package.json
@@ -31,7 +31,6 @@
     "next": "15.1.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-markdown": "^9.0.3",
     "react-syntax-highlighter": "^15.6.1",
     "rehype-sanitize": "^6.0.0",
     "remark-gfm": "^4.0.0",
