+++ src/features/codingChallenges/content/exercises/reduce-max.mdx
@@ -1,0 +1,98 @@
+"use client"
+
+export const metadata = {
+  slug: "reduce-max",
+  title: "Find Maximum with reduce()",
+  category: { name: "array", method: "reduce" } as const,
+  starterCode: `const solve = (numbers) => {
+  // Handle empty array case first
+  if (numbers.length === 0) return null;
+
+  // Use reduce to find maximum
+  return numbers[0]; // Replace with your logic
+}`,
+  testCases: [
+    { input: [[5, 2, 9, 1]], expected: 9, message: "Should find maximum `9` in `[5, 2, 9, 1]`" },
+    { input: [[1, 2, 3, 4, 5]], expected: 5, message: "Should find maximum `5` in `[1, 2, 3, 4, 5]`" },
+    { input: [[]], expected: null, message: "Should return `null` for empty array `[]`" },
+    { input: [[42]], expected: 42, message: "Should return `42` for single-element array `[42]`" },
+  ],
+  excerpt: "Use `reduce` to find the largest number in an array.",
+};
+
+# Find Maximum with reduce()
+
+### Problem
+Write a function that uses `reduce` to find the largest number in an array.
+
+### Example
+```js
+Input: [5, 2, 9, 1]
+Output: 9
+```
+
+### Requirements
+1. Use Array.prototype.reduce()
+2. If the array is empty, return null
+3. If there's only one element, return that element
+
+## Finding Maximum with reduce()
+
+### Core Concept
+`reduce()` can track and update a "running maximum" as it processes each array element.
+
+### How It Works
+```js
+array.reduce((max, current) => {
+  return current > max ? current : max;
+});
+```
+
+### Key Points
+1. **No Initial Value Needed**
+   - First element becomes initial maximum
+   - Perfect for number comparisons
+
+2. **Comparison Logic**
+   - Each iteration compares current vs max
+   - Larger value becomes new maximum
+
+3. **Edge Cases**
+   - Empty arrays need special handling
+   - Single element is automatically maximum
+
+### Why reduce() for Maximum?
+- Single pass through array (O(n))
+- Clean, functional approach
+- Built-in array bounds handling
+
+## Visual Walkthrough
+```js
+[5, 2, 9, 1].reduce((max, current) => Math.max(max, current))
+```
+
+### Process:
+```
+Start: max = 5 (first element)
+
+Compare with 2:
+   max = Math.max(5, 2) = 5
+
+Compare with 9:
+   max = Math.max(5, 9) = 9
+
+Compare with 1:
+   max = Math.max(9, 1) = 9
+
+Result: 9
+```
+
+### Common Mistakes
+- **Wrong Initial**: Using `0` as initialValue (fails with negative numbers)
+- **Missing Null**: Not handling empty array case properly
+- **Complex Logic**: Overthinking the comparison (use `Math.max`)
+
+### Tips
+- **Use Math.max**: Cleaner than manual comparison
+- **Type Check**: Validate input array contains numbers
+- **Edge Cases**: Handle empty and single-element arrays first
